SvelteKit routes auto migration
