+++ package.json
@@ -23,41 +23,41 @@
     "@iconify-json/fa-brands": "^1.1.2",
     "@iconify-json/fa-solid": "^1.1.2",
     "@iconify-json/heroicons-solid": "^1.1.2",
-    "@iconify-json/ic": "^1.1.7",
+    "@iconify-json/ic": "^1.1.9",
     "@iconify-json/ion": "^1.1.2",
-    "@iconify-json/mdi": "^1.1.19",
-    "@iconify-json/octicon": "^1.1.14",
+    "@iconify-json/mdi": "^1.1.30",
+    "@iconify-json/octicon": "^1.1.16",
     "@iconify-json/ri": "^1.1.2",
-    "@iconify-json/simple-icons": "^1.1.20",
+    "@iconify-json/simple-icons": "^1.1.22",
     "@mapbox/mapbox-gl-geocoder": "^5.0.1",
-    "@playwright/test": "^1.24.2",
+    "@playwright/test": "^1.25.0",
     "@rollup/plugin-yaml": "^3.1.0",
     "@sveltejs/adapter-static": "^1.0.0-next.39",
-    "@sveltejs/kit": "^1.0.0-next.405",
+    "@sveltejs/kit": "^1.0.0-next.428",
     "@sveltejs/vite-plugin-svelte": "^1.0.1",
     "@types/events": "^3.0.0",
     "@types/mapbox-gl": "^2.7.4",
-    "@types/mapbox__mapbox-gl-geocoder": "^4.7.2",
-    "@typescript-eslint/eslint-plugin": "^5.31.0",
-    "@typescript-eslint/parser": "^5.31.0",
-    "eslint": "^8.17.0",
+    "@types/mapbox__mapbox-gl-geocoder": "^4.7.3",
+    "@typescript-eslint/eslint-plugin": "^5.33.1",
+    "@typescript-eslint/parser": "^5.33.1",
+    "eslint": "^8.22.0",
     "eslint-plugin-svelte3": "^4.0.0",
     "js-yaml": "^4.1.0",
     "mapbox-gl": "^2.9.2",
     "marked": "^4.0.18",
-    "playwright": "^1.24.1",
+    "playwright": "^1.25.0",
     "prettier": "^2.7.1",
     "prettier-plugin-svelte": "^2.7.0",
     "svelte": "3.49.0",
-    "svelte-check": "^2.7.2",
+    "svelte-check": "^2.8.1",
     "svelte-color-mode": "^0.1.4",
     "svelte-multiselect": "^5.0.4",
     "svelte-preprocess": "^4.10.7",
     "svelte-range-slider-pips": "^2.0.3",
     "svelte-toc": "^0.2.10",
-    "svelte2tsx": "^0.5.13",
+    "svelte2tsx": "^0.5.14",
     "typescript": "^4.7.3",
     "unplugin-icons": "^0.14.7",
-    "vite": "^3.0.4"
+    "vite": "^3.0.9"
   }
 }
