+++ src/routes/auszeichnungen/+page.ts
@@ -1,0 +1,9 @@
+import type { PageLoad } from '@sveltejs/kit'
+import { fetchPage, fetchYamlList } from '../../fetch'
+
+export const load: PageLoad = async () => {
+  const page = await fetchPage(`auszeichnungen`)
+  const awards = await fetchYamlList(`Auszeichnungen`, `auszeichnungen#`)
+
+  return { page, awards }
+}
