+++ src/routes/blog/+page.svelte
@@ -1,46 +1,25 @@
-<script lang="ts" context="module">
+<script lang="ts">
   import PostPreview from '$src/components/PostPreview.svelte'
   import Social from '$src/components/Social.svelte'
   import TagList from '$src/components/TagList.svelte'
-  import { fetchPosts, fetchYaml } from '$src/fetch'
-  import type { BlogTag, Post } from '$src/types'
-  import type { Load } from '@sveltejs/kit'
+  import type { BlogTag } from '$src/types'
   import { flip } from 'svelte/animate'
   import { scale } from 'svelte/transition'
+  import type { PageData } from './$types'
 
-  export const load: Load = async () => {
-    const posts = await fetchPosts()
-    const social = await fetchYaml(`Social`)
-
-    return { props: { posts, social } }
-  }
-</script>
-
-<script lang="ts">
-  export let posts: Post[]
-  export let social: Record<string, string>
+  export let data: PageData
+  $: ({ posts, social } = data)
 
   let activeTag: BlogTag
 
   $: filteredPosts = posts.filter(
     (post) => activeTag === `Alle` || post.tags.includes(activeTag)
   )
-
-  const tagCounter: Record<string, number> = { Alle: posts.length }
-
-  // count tag occurrences
-  for (const post of posts) {
-    for (const tag of post.tags) {
-      tagCounter[tag] = (tagCounter[tag] ?? 0) + 1
-    }
-  }
-
-  const tagOccurrences = Object.entries(tagCounter) as [BlogTag, number][]
 </script>
 
 <Social {social} fixed vertical />
 
-<TagList {tagOccurrences} bind:activeTag />
+<TagList {posts} bind:activeTag />
 
 <ul>
   {#each filteredPosts as post (post.slug)}
