+++ src/routes/blog/[slug]/+page.ts
@@ -1,0 +1,11 @@
+import { fetchPost } from '$src/fetch'
+import type { PageLoad } from '@sveltejs/kit'
+import { error } from '@sveltejs/kit'
+
+export const load: PageLoad = async ({ params }) => {
+  const post = await fetchPost(params.slug)
+
+  if (!post) throw error(404)
+
+  return { post }
+}
