+++ src/routes/faq/+page.ts
@@ -1,0 +1,7 @@
+import type { PageLoad } from '@sveltejs/kit'
+import { fetchYamlList } from '../../fetch'
+
+export const load: PageLoad = async () => {
+  const faqs = await fetchYamlList(`FAQ`, `faq#`)
+  return { faqs }
+}
